requirements.txt
# Core dependencies
neo4j>=5.15.0
python-dotenv>=1.0.0

# Data processing
polars>=1.0.0
pandas>=2.0.0

# Embeddings and ML
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0

# Vector store
faiss-cpu>=1.7.4
numpy>=1.24.0

# LLM
anthropic>=0.25.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
typing-extensions>=4.8.0

# Utilities
tqdm>=4.65.0
pydantic>=2.0.0

# Graph analysis
networkx>=3.0
leidenalg>=0.9.0
python-igraph>=0.11.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
